+++ plugins/kubernetes/src/components/KubernetesContent.tsx
@@ -24,6 +24,7 @@ import { Cluster } from './Cluster';
 import EmptyStateImage from '../assets/emptystate.svg';
 import { useKubernetesObjects } from '../hooks';
 import { Content, Page, Progress } from '@backstage/core-components';
+import { DetectedErrorsContext } from '../hooks/useMatchingErrors';
 
 type KubernetesContentProps = {
   entity: Entity;
@@ -49,88 +50,92 @@ export const KubernetesContent = ({
       : new Map<string, DetectedError[]>();
 
   return (
-    <Page themeId="tool">
-      <Content>
-        {kubernetesObjects === undefined && error === undefined && <Progress />}
+    <DetectedErrorsContext.Provider value={[...detectedErrors.values()].flat()}>
+      <Page themeId="tool">
+        <Content>
+          {kubernetesObjects === undefined && error === undefined && (
+            <Progress />
+          )}
 
-        {/* errors retrieved from the kubernetes clusters */}
-        {clustersWithErrors.length > 0 && (
-          <Grid container spacing={3} direction="column">
-            <Grid item>
-              <ErrorPanel
-                entityName={entity.metadata.name}
-                clustersWithErrors={clustersWithErrors}
-              />
+          {/* errors retrieved from the kubernetes clusters */}
+          {clustersWithErrors.length > 0 && (
+            <Grid container spacing={3} direction="column">
+              <Grid item>
+                <ErrorPanel
+                  entityName={entity.metadata.name}
+                  clustersWithErrors={clustersWithErrors}
+                />
+              </Grid>
             </Grid>
-          </Grid>
-        )}
+          )}
 
-        {/* other errors */}
-        {error !== undefined && (
-          <Grid container spacing={3} direction="column">
-            <Grid item>
-              <ErrorPanel
-                entityName={entity.metadata.name}
-                errorMessage={error}
-              />
+          {/* other errors */}
+          {error !== undefined && (
+            <Grid container spacing={3} direction="column">
+              <Grid item>
+                <ErrorPanel
+                  entityName={entity.metadata.name}
+                  errorMessage={error}
+                />
+              </Grid>
             </Grid>
-          </Grid>
-        )}
+          )}
 
-        {kubernetesObjects && (
-          <Grid container spacing={3} direction="column">
-            <Grid item>
-              <ErrorReporting detectedErrors={detectedErrors} />
-            </Grid>
-            <Grid item>
-              <Typography variant="h3">Your Clusters</Typography>
-            </Grid>
-            <Grid item container>
-              {kubernetesObjects?.items.length <= 0 && (
-                <Grid
-                  container
-                  justifyContent="space-around"
-                  direction="row"
-                  alignItems="center"
-                  spacing={2}
-                >
-                  <Grid item xs={4}>
-                    <Typography variant="h5">
-                      No resources on any known clusters for{' '}
-                      {entity.metadata.name}
-                    </Typography>
-                  </Grid>
-                  <Grid item xs={4}>
-                    <img
-                      src={EmptyStateImage}
-                      alt="EmptyState"
-                      data-testid="emptyStateImg"
-                    />
-                  </Grid>
-                </Grid>
-              )}
-              {kubernetesObjects?.items.length > 0 &&
-                kubernetesObjects?.items.map((item, i) => {
-                  const podsWithErrors = new Set<string>(
-                    detectedErrors
-                      .get(item.cluster.name)
-                      ?.filter(de => de.sourceRef.kind === 'Pod')
-                      .map(de => de.sourceRef.name),
-                  );
-
-                  return (
-                    <Grid item key={i} xs={12}>
-                      <Cluster
-                        clusterObjects={item}
-                        podsWithErrors={podsWithErrors}
+          {kubernetesObjects && (
+            <Grid container spacing={3} direction="column">
+              <Grid item>
+                <ErrorReporting detectedErrors={detectedErrors} />
+              </Grid>
+              <Grid item>
+                <Typography variant="h3">Your Clusters</Typography>
+              </Grid>
+              <Grid item container>
+                {kubernetesObjects?.items.length <= 0 && (
+                  <Grid
+                    container
+                    justifyContent="space-around"
+                    direction="row"
+                    alignItems="center"
+                    spacing={2}
+                  >
+                    <Grid item xs={4}>
+                      <Typography variant="h5">
+                        No resources on any known clusters for{' '}
+                        {entity.metadata.name}
+                      </Typography>
+                    </Grid>
+                    <Grid item xs={4}>
+                      <img
+                        src={EmptyStateImage}
+                        alt="EmptyState"
+                        data-testid="emptyStateImg"
                       />
                     </Grid>
-                  );
-                })}
+                  </Grid>
+                )}
+                {kubernetesObjects?.items.length > 0 &&
+                  kubernetesObjects?.items.map((item, i) => {
+                    const podsWithErrors = new Set<string>(
+                      detectedErrors
+                        .get(item.cluster.name)
+                        ?.filter(de => de.sourceRef.kind === 'Pod')
+                        .map(de => de.sourceRef.name),
+                    );
+
+                    return (
+                      <Grid item key={i} xs={12}>
+                        <Cluster
+                          clusterObjects={item}
+                          podsWithErrors={podsWithErrors}
+                        />
+                      </Grid>
+                    );
+                  })}
+              </Grid>
             </Grid>
-          </Grid>
-        )}
-      </Content>
-    </Page>
+          )}
+        </Content>
+      </Page>
+    </DetectedErrorsContext.Provider>
   );
 };