+++ packages/backend-app-api/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/backend-app-api
 
+## 0.4.3
+
+### Patch Changes
+
+- cf13b482f9e: Switch `configServiceFactory` to use `ConfigSources` from `@backstage/config-loader` to load config.
+- Updated dependencies
+  - @backstage/backend-common@0.18.5
+  - @backstage/config-loader@1.3.0
+  - @backstage/plugin-permission-node@0.7.8
+  - @backstage/backend-tasks@0.5.2
+  - @backstage/plugin-auth-node@0.2.14
+  - @backstage/backend-plugin-api@0.5.2
+  - @backstage/cli-common@0.1.12
+  - @backstage/config@1.0.7
+  - @backstage/errors@1.1.5
+  - @backstage/types@1.0.2
+
 ## 0.4.3-next.1
 
 ### Patch Changes