+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/test-utils
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/theme@0.3.0
+  - @backstage/core-app-api@1.8.0
+  - @backstage/config@1.0.7
+  - @backstage/core-plugin-api@1.5.1
+  - @backstage/types@1.0.2
+  - @backstage/plugin-permission-common@0.7.5
+  - @backstage/plugin-permission-react@0.4.12
+
 ## 1.3.1-next.2
 
 ### Patch Changes