+++ plugins/adr-backend/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-adr-backend
 
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.18.5
+  - @backstage/integration@1.4.5
+  - @backstage/catalog-client@1.4.1
+  - @backstage/catalog-model@1.3.0
+  - @backstage/config@1.0.7
+  - @backstage/errors@1.1.5
+  - @backstage/plugin-adr-common@0.2.9
+  - @backstage/plugin-search-common@1.2.3
+
 ## 0.3.3-next.1
 
 ### Patch Changes