+++ plugins/auth-backend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-auth-backend
 
+## 0.18.3
+
+### Patch Changes
+
+- 7c116bcac7f: Fixed the way that some request errors are thrown
+- 473db605a4f: Fix config schema definition.
+- 3ffcdac7d07: Added a persistent session store through the database
+- Updated dependencies
+  - @backstage/backend-common@0.18.5
+  - @backstage/plugin-auth-node@0.2.14
+  - @backstage/catalog-client@1.4.1
+  - @backstage/catalog-model@1.3.0
+  - @backstage/config@1.0.7
+  - @backstage/errors@1.1.5
+  - @backstage/types@1.0.2
+
 ## 0.18.3-next.2
 
 ### Patch Changes