+++ plugins/events-backend-module-github/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-events-backend-module-github
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.5.2
+  - @backstage/config@1.0.7
+  - @backstage/plugin-events-node@0.2.6
+
 ## 0.1.7-next.1
 
 ### Patch Changes