+++ plugins/sonarqube-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-sonarqube-backend
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.18.5
+  - @backstage/config@1.0.7
+  - @backstage/errors@1.1.5
+
 ## 0.1.10-next.1
 
 ### Patch Changes