+++ plugins/techdocs-module-addons-contrib/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-techdocs-module-addons-contrib
 
+## 1.0.13
+
+### Patch Changes
+
+- 6afc7f052ca: Show cursor pointer when hovering on lightbox
+- Updated dependencies
+  - @backstage/theme@0.3.0
+  - @backstage/integration@1.4.5
+  - @backstage/core-components@0.13.1
+  - @backstage/integration-react@1.1.13
+  - @backstage/plugin-techdocs-react@1.1.6
+  - @backstage/core-plugin-api@1.5.1
+
 ## 1.0.13-next.2
 
 ### Patch Changes